Create FOREX Master Pattern Trend Finder by nnamdert
What does this Indicator do?

This indicator works by identifying Trends and "potential" shifts in trends before they happen. It is technically a momentum indicator that watches for specific Bullish or Bearish momentum (i.e. Bull Runs, and Crashes).

How does it do this?

It uses a custom script that takes into account extremes in price and certain moving averages and other accurate forward looking indicators. Once it finds a Bull run, signals are triggered in the form of PLUS signs. It does the same for strong Bearish trends. Based on how the script is configured, the Indicator is able to spot (in advance) unique movement in the trend that can predict a change in momentum prior to the change occurring.

These momentum changes can be configured as alerts.