Create GRIDBOT Scalper
What is this Indicator used for?
Made specifically for GRID Bots

note: before continuing... this indicator works on any timeframe, but it WORKS BEST ON THE 15 MINUTE TIMEFRAME

Straters and Forex Master Pattern Value Line Traders use this to help determine when the price could reverse.

This indicator is a scalping indicator that produces signals when a "potential" reversal in price is indicated. When the price moves UP and a Potential Bearish Reversal Signal occurs, traders can use this signal as a potential SHORT entry signal for their Short Grid Bot. The process is the same in reverse. After a sustained move down, a Potential Bullish Signal can be used by the trader as a potential LONG entry signal for their GridBot.