Create Simple STRAT Tool by nnam
What this Indicator Does

This indicator is a very simple tool created specifically for experienced Straters. It was created for those Straters who fully understand the 1-2-3 Strat Scenarios, are in need of an easy to use tool, and do not want or need a lot of messy markings on their chart.

The indicator simply allows the user to color code the Strat 1, 2 ,3 (Inside /Outside /Up / Down) Bars as desired and by default extends lines to the right of the chart from the Highs and Lows of the previous 2 Bars giving the user a simple reference for Strat scenario structure breaks.