Create Expansion Finder by nnam
What this Indicator Does

This indicator helps the trader locate expansion and contraction areas in an easy visual way.

When the asset moves from a contraction phase into an expansion phase, the bars change color (customizable). This allows the trader to recognize areas of contraction and avoid trading them. Once a Bar Range moves outside of the average range as specified by the user, the bar will change color informing the trader that the current bar and by default the market, is moving into an expansion phase from a contraction phase.

The indicator works well for those traders that use the Forex Master Pattern to locate Value Lines and Value Areas on the chart giving them an opportunity to draw in these areas with ease.