Create Easy RSI by nnamdert
What Does this Indicator Do?

The Easy RSI Indicator color codes candles based on their RSI Value vs. Open / Close (Red / Green). It plots the current price and current RSI value on the chart in real-time. Additionally, when the RSI Value is in an oversold or overbought condition, it plots that signal on the chart in real-time.

The initial candle color is the standard Red / Green Tradingview color, but a Gradient is added to the color which either darkens or lightens the color based on the RSI Value.

The higher the RSI Value, the brighter the Green Color is. The lower the RSI Value, the brighter the Red Color is.