Create Contraction Finder by nnamdert
What is this indicator?
An indicator that finds Highs and Lows and plots them as lines on the chart.

What does it do?
Simple Indicator that finds Highest High within a specified lookback period. Then, finds the lowest lows since then.
Indicator then looks for the Lowest Low within a shorter loopback period and finds the Highest High since then.
Symmetrical Triangle are plotted as are trend lines if used as a contraction finder.